fix lint

+++ crates/scroll/alloy/rpc-types/src/receipt.rs
@@ -45,7 +45,7 @@ impl alloy_network_primitives::ReceiptResponse for ScrollTransactionReceipt {
         self.inner.transaction_index()
     }
 
-    fn gas_used(&self) -> u128 {
+    fn gas_used(&self) -> u64 {
         self.inner.gas_used()
     }
 
@@ -53,7 +53,7 @@ impl alloy_network_primitives::ReceiptResponse for ScrollTransactionReceipt {
         self.inner.effective_gas_price()
     }
 
-    fn blob_gas_used(&self) -> Option<u128> {
+    fn blob_gas_used(&self) -> Option<u64> {
         self.inner.blob_gas_used()
     }
 
@@ -73,7 +73,7 @@ impl alloy_network_primitives::ReceiptResponse for ScrollTransactionReceipt {
         self.inner.authorization_list()
     }
 
-    fn cumulative_gas_used(&self) -> u128 {
+    fn cumulative_gas_used(&self) -> u64 {
         self.inner.cumulative_gas_used()
     }
